+++ src/chipeitlib/src/main/java/es/chipeit/lib/interfaces/IRegisters.kt
@@ -3,7 +3,7 @@ package es.chipeit.lib.interfaces
 internal interface IRegisters {
     val v: IMemory<Byte>
     var i: Short
-    var pc: Short
+    var pc: Int
 
     var sp: Byte
 }
